Move RootMetadataSigned to kbfsmd
Make libkbfs.RootMetadataSigned wrap kbfsmd.RootMetadataSigned.
