+++ libkbfs/reporter_kbpki.go
@@ -151,7 +151,7 @@ func (r *ReporterKBPKI) ReportErr(ctx context.Context,
 	case DirTooBigError:
 		code = keybase1.FSErrorType_NOT_IMPLEMENTED
 		params[errorParamFeature] = errorFeatureDirLimit
-	case NewMetadataVersionError:
+	case kbfsmd.NewMetadataVersionError:
 		code = keybase1.FSErrorType_OLD_VERSION
 		err = OutdatedVersionError{}
 	case NewDataVersionError:
