+++ libkbfs/root_metadata_test.go
@@ -110,62 +110,6 @@ func runBenchmarkOverMetadataVers(
 // TODO: Add way to test with all possible (ver, maxVer) combos,
 // e.g. for upconversion tests.
 
-func testRootMetadataFinalVerify(t *testing.T, ver MetadataVer) {
-	tlfID := tlf.FakeID(1, tlf.Private)
-
-	uid := keybase1.MakeTestUID(1)
-	bh, err := tlf.MakeHandle(
-		[]keybase1.UserOrTeamID{uid.AsUserOrTeam()}, nil, nil, nil, nil)
-	require.NoError(t, err)
-
-	brmd, err := kbfsmd.MakeInitialRootMetadata(ver, tlfID, bh)
-	require.NoError(t, err)
-
-	ctx := context.Background()
-	codec := kbfscodec.NewMsgpack()
-	signer := kbfscrypto.SigningKeySigner{
-		Key: kbfscrypto.MakeFakeSigningKeyOrBust("key"),
-	}
-
-	extra := kbfsmd.FakeInitialRekey(brmd, bh, kbfscrypto.TLFPublicKey{})
-
-	brmd.SetLastModifyingWriter(uid)
-	brmd.SetLastModifyingUser(uid)
-	brmd.SetSerializedPrivateMetadata([]byte{42})
-	err = brmd.SignWriterMetadataInternally(ctx, codec, signer)
-	require.NoError(t, err)
-
-	rmds, err := SignBareRootMetadata(
-		ctx, codec, signer, signer, brmd, time.Time{})
-	require.NoError(t, err)
-
-	// verify it
-	err = rmds.IsValidAndSigned(ctx, codec, nil, extra)
-	require.NoError(t, err)
-
-	ext, err := tlf.NewHandleExtension(
-		tlf.HandleExtensionFinalized, 1, "fake user", time.Now())
-	require.NoError(t, err)
-
-	// make a final copy
-	rmds2, err := rmds.MakeFinalCopy(codec, time.Now(), ext)
-	require.NoError(t, err)
-
-	// verify the finalized copy
-	err = rmds2.IsValidAndSigned(ctx, codec, nil, extra)
-	require.NoError(t, err)
-
-	// touch something the server shouldn't be allowed to edit for
-	// finalized metadata and verify verification failure.
-	md3, err := rmds2.MD.DeepCopy(codec)
-	require.NoError(t, err)
-	md3.SetRekeyBit()
-	rmds3 := rmds2
-	rmds2.MD = md3
-	err = rmds3.IsValidAndSigned(ctx, codec, nil, extra)
-	require.NotNil(t, err)
-}
-
 type privateMetadataFuture struct {
 	PrivateMetadata
 	Dir dirEntryFuture
@@ -979,7 +923,6 @@ func TestRootMetadataTeamMakeSuccessor(t *testing.T) {
 
 func TestRootMetadata(t *testing.T) {
 	tests := []func(*testing.T, MetadataVer){
-		testRootMetadataFinalVerify,
 		testRootMetadataGetTlfHandlePublic,
 		testRootMetadataGetTlfHandlePrivate,
 		testRootMetadataLatestKeyGenerationPrivate,
